dev-requirements.txt
filetype==1.2.0
mutagen==1.47.0
numpy==1.26.2
Pillow==10.1.0
pydub==0.25.1
soundfile==0.12.1
git+https://github.com/ego-lay-atman-bay/audioman.git
